Refactor and cleanup Flywheel and Funnel, add CANConstants
