+++ src/main/java/com/team1165/robot/subsystems/roller/flywheel/FlywheelConstants.java
@@ -5,7 +5,7 @@
  * the root directory of this project.
  */
 
-package com.team1165.robot.subsystems.roller.flywheel.constants;
+package com.team1165.robot.subsystems.roller.flywheel;
 
 import static edu.wpi.first.units.Units.KilogramSquareMeters;
 import static edu.wpi.first.units.Units.Volts;
@@ -14,25 +14,29 @@ import com.revrobotics.spark.SparkLowLevel.MotorType;
 import com.revrobotics.spark.config.SparkBaseConfig;
 import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
 import com.revrobotics.spark.config.SparkMaxConfig;
+import com.team1165.robot.globalconstants.CANConstants.IDs.RIO;
 import com.team1165.robot.util.vendor.rev.SparkConfig;
 import edu.wpi.first.math.system.plant.DCMotor;
 import org.ironmaple.simulation.motorsims.SimMotorConfigs;
 
-/** Class to store constants for the Flywheel subsystem of the robot. */
+/** Class to store constants for the {@link Flywheel} subsystem of the robot. */
 public class FlywheelConstants {
-  /** Class to store constants required for initial configuration of the Flywheel. */
+  /** Class to store constants required for initial configuration of the Flywheel subsystem. */
   public static final class Configurations {
-    // SPARK MAX configurations
+    // Base SPARK configuration
     private static final SparkBaseConfig baseMotorConfig =
         new SparkMaxConfig().smartCurrentLimit(50).idleMode(IdleMode.kBrake);
+
+    // Individual SPARK MAX configurations
     public static final SparkConfig primaryMotorConfig =
-        SparkConfig.sparkMax("FlywheelPrimary", 1, MotorType.kBrushless, baseMotorConfig);
+        SparkConfig.sparkMax(
+            "FlywheelPrimary", RIO.flywheelPrimary, MotorType.kBrushless, baseMotorConfig);
     public static final SparkConfig secondaryMotorConfig =
         SparkConfig.sparkMax(
             "FlywheelSecondary",
-            2,
+            RIO.funnelSecondary,
             MotorType.kBrushless,
-            baseMotorConfig); // TODO: Likely will need to invert this motor?
+            baseMotorConfig.inverted(true));
 
     // Simulation motor configuration
     public static final SimMotorConfigs simConfig =
